requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argcomplete==3.3.0
attrs==23.2.0
awscli
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.34.40
botocore==1.34.67
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click
cloudpathlib==0.16.0
# colorama  # Removed
confection==0.1.4
configobj==5.0.8
configparser==6.0.0
ctransformers==0.2.27
cymem==2.0.8
dataclasses-json==0.6.4
distlib==0.3.8
distro==1.9.0
dnspython==2.5.0
docutils==0.16
email-validator==2.1.0.post1
etelemetry==0.3.1
fastapi==0.109.2
fastavro==1.9.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langcodes==3.3.0
langsmith==0.0.87
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
mistralai
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.4
openai==1.12.0
orjson==3.9.13
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
pillow==10.2.0
pipx
platformdirs==4.2.0
preshed==3.0.9
prov==2.0.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
pydot==2.0.0
pyenv-win==3.1.1
PyMuPDFb==1.23.22
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
rsa==4.7.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.1.1
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.26
srsly==2.4.8
starlette==0.36.3
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.2.0
tqdm
traits==6.3.2
transformers==4.37.2
typer==0.9.0
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.0.7
userpath==1.9.2
uuid==1.30
uvicorn==0.27.1
virtualenv==20.25.1
wasabi
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
XlsxWriter==3.2.0
yarl==1.9.4
